Connection Protocol: Address review comments
