+++ src/lib/protocols/connections/domain/Connection.ts
@@ -74,8 +74,8 @@ export class Connection extends EventEmitter {
         if (newState == state) {
           resolve(true);
         }
-      })
-    })
+      });
+    });
   }
 
   async isConnected() {
